Update the attribute annotation package and updates the input annotations.
