+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/parser/DefaultParser.java
@@ -20,7 +20,7 @@
 package org.orbisgis.orbistoolbox.controller.parser;
 
 import org.orbisgis.orbistoolbox.model.*;
-import org.orbisgis.wpsgroovyapi.model.InputAttribute;
+import org.orbisgis.wpsgroovyapi.attributes.InputAttribute;
 import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Field;
