+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/parser/ParserController.java
@@ -24,8 +24,8 @@ import groovy.lang.GroovyRuntimeException;
 import groovy.lang.GroovyShell;
 import org.orbisgis.orbistoolbox.model.*;
 import org.orbisgis.orbistoolbox.model.Process;
-import org.orbisgis.wpsgroovyapi.model.InputAttribute;
-import org.orbisgis.wpsgroovyapi.model.OutputAttribute;
+import org.orbisgis.wpsgroovyapi.attributes.InputAttribute;
+import org.orbisgis.wpsgroovyapi.attributes.OutputAttribute;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
