+++ bundles/orbistoolbox/src/main/java/org/orbisgis/orbistoolbox/controller/parser/RawDataParser.java
@@ -20,7 +20,7 @@
 package org.orbisgis.orbistoolbox.controller.parser;
 
 import org.orbisgis.orbistoolbox.model.*;
-import org.orbisgis.wpsgroovyapi.model.*;
+import org.orbisgis.wpsgroovyapi.attributes.*;
 import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Field;
