+++ bundles/wps-groovy-api/src/main/java/org/orbisgis/wpsgroovyapi/attributes/OutputAttribute.groovy
@@ -27,33 +27,24 @@
  * info_at_ orbisgis.org
  */
 
-package org.orbisgis.wpsgroovyapi.model
+package org.orbisgis.wpsgroovyapi.attributes
 
 import java.lang.annotation.Retention
 import java.lang.annotation.RetentionPolicy
 
 /**
- * Attributes for the BoundingBox.
- * Bounding box data serves a variety of purposes in spatial data processing.
- * Some simple applications are the definition of extents for a clipping operation or the definition of an
- * analysis region.
- *
- * The following fields must be defined (mandatory) :
- *  - format : FormatAttribute[]
- *      Identifies a valid format for an input or output.
+ * Attributes for the Output.
+ * The output is the process return value.
  *
  * The following fields can be defined (optional) :
- *  - supportedCRSList : SupportedCRSAttribute[]
- *      List of CRS supported by the BoundingBox data.
+ *  - output : OutputAttribute[]
+ *      Nested Output.
  *
  * @author Sylvain PALOMINOS
  */
 @Retention(RetentionPolicy.RUNTIME)
-@interface BoundingBoxAttribute {
+@interface OutputAttribute {
 
-    /** Identifies a valid format for an input or output. */
-    FormatAttribute[] format()
-
-    /** List of CRS supported by the BoundingBox data. */
-    SupportedCRSAttribute[] supportedCRSList()
+    /** Nested Output. */
+    OutputAttribute[] output() default []
 }
