+++ bundles/wps-groovy-api/src/main/java/org/orbisgis/wpsgroovyapi/input/EnumerationInput.groovy
@@ -1,35 +1,78 @@
 /**
- * OrbisToolBox is an OrbisGIS plugin dedicated to create and manage processing.
+ * OrbisGIS is a GIS application dedicated to scientific spatial analysis.
+ * This cross-platform GIS is developed at the Lab-STICC laboratory by the DECIDE
+ * team located in University of South Brittany, Vannes.
  *
- * OrbisToolBox is distributed under GPL 3 license. It is produced by CNRS <http://www.cnrs.fr/> as part of the
- * MApUCE project, funded by the French Agence Nationale de la Recherche (ANR) under contract ANR-13-VBDU-0004.
+ * OrbisGIS is distributed under GPL 3 license.
  *
- * OrbisToolBox is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
- * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any
- * later version.
+ * Copyright (C) 2007-2014 IRSTV (FR CNRS 2488)
+ * Copyright (C) 2015-2016 CNRS (UMR CNRS 6285)
  *
- * OrbisToolBox is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ * This file is part of OrbisGIS.
  *
- * You should have received a copy of the GNU General Public License along with OrbisToolBox. If not, see
- * <http://www.gnu.org/licenses/>.
+ * OrbisGIS is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later
+ * version.
  *
- * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
+ * OrbisGIS is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * OrbisGIS. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * For more information, please consult: <http://www.orbisgis.org/>
+ * or contact directly:
+ * info_at_ orbisgis.org
  */
 
 package org.orbisgis.wpsgroovyapi.input
 
 import groovy.transform.AnnotationCollector
 import groovy.transform.Field
-import org.orbisgis.wpsgroovyapi.model.EnumerationAttribute
-import org.orbisgis.wpsgroovyapi.model.DescriptionTypeAttribute
-import org.orbisgis.wpsgroovyapi.model.InputAttribute
+import org.orbisgis.wpsgroovyapi.attributes.EnumerationAttribute
+import org.orbisgis.wpsgroovyapi.attributes.DescriptionTypeAttribute
+import org.orbisgis.wpsgroovyapi.attributes.InputAttribute
 
 /**
- * The Enumeration input allow the user to select one or more value from a predefined list.
+ * Enumeration input annotation.
+ * The Enumeration complex data represents a selection of values from a predefined list.
+ * As an input, this annotation should be placed just before the variable.
+ *
+ * The following fields must be defined (mandatory) :
+ *  - title : String
+ *       Title of a process, input, and output. Normally available for display to a human.
+ *  - values : String[]
+ *      List of possible values.
+ *
+ * The following fields can be defined (optional) :
+ *  - resume : String
+ *      Brief narrative description of a process, input, and output. Normally available for display to a human..
+ *  - keywords : String
+ *      Coma separated keywords that characterize a process, its inputs, and outputs.
+ *  - identifier : String
+ *      Unambiguous identifier of a process, input, and output. It should be a valid URI.
+ *  - metadata : MetaData[]
+ *      Reference to additional metadata about this item.
+ *  - minOccurs : int
+ *      Minimum number of times that values for this parameter are required. 0 means the input is optional.
+ *  - maxOccurs : int
+ *      Maximum number of times that this parameter may be present.
+ *  - input : InputAttribute[]
+ *      Nested Input.
+ *  - multiSelection : boolean
+ *      Allow or not to select more than one value.
+ *  - isEditable : boolean
+ *      Enable or not the user to use its own value.
+ *  - names : String[]
+ *      Displayable name of the values. If not specified, use the values as name.
+ *  - defaultValues : String[]
+ *      Default selected values, can be empty.
+ *
+ * Usage example can be found at https://github.com/orbisgis/orbisgis/wiki/
  *
  * @author Sylvain PALOMINOS
  */
 @AnnotationCollector([Field, EnumerationAttribute, InputAttribute, DescriptionTypeAttribute])
-@interface EnumerationInput {
-}
+@interface EnumerationInput {}
