+++ bundles/wps-groovy-api/src/main/java/org/orbisgis/wpsgroovyapi/input/LiteralDataInput.groovy
@@ -1,36 +1,74 @@
 /**
- * OrbisToolBox is an OrbisGIS plugin dedicated to create and manage processing.
+ * OrbisGIS is a GIS application dedicated to scientific spatial analysis.
+ * This cross-platform GIS is developed at the Lab-STICC laboratory by the DECIDE
+ * team located in University of South Brittany, Vannes.
  *
- * OrbisToolBox is distributed under GPL 3 license. It is produced by CNRS <http://www.cnrs.fr/> as part of the
- * MApUCE project, funded by the French Agence Nationale de la Recherche (ANR) under contract ANR-13-VBDU-0004.
+ * OrbisGIS is distributed under GPL 3 license.
  *
- * OrbisToolBox is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
- * License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any
- * later version.
+ * Copyright (C) 2007-2014 IRSTV (FR CNRS 2488)
+ * Copyright (C) 2015-2016 CNRS (UMR CNRS 6285)
  *
- * OrbisToolBox is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ * This file is part of OrbisGIS.
  *
- * You should have received a copy of the GNU General Public License along with OrbisToolBox. If not, see
- * <http://www.gnu.org/licenses/>.
+ * OrbisGIS is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later
+ * version.
  *
- * For more information, please consult: <http://www.orbisgis.org/> or contact directly: info_at_orbisgis.org
+ * OrbisGIS is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * OrbisGIS. If not, see <http://www.gnu.org/licenses/>.
+ *
+ * For more information, please consult: <http://www.orbisgis.org/>
+ * or contact directly:
+ * info_at_ orbisgis.org
  */
 
 package org.orbisgis.wpsgroovyapi.input
 
 import groovy.transform.AnnotationCollector
 import groovy.transform.Field
-import org.orbisgis.wpsgroovyapi.model.DescriptionTypeAttribute
-import org.orbisgis.wpsgroovyapi.model.InputAttribute
-import org.orbisgis.wpsgroovyapi.model.LiteralDataAttribute
+import org.orbisgis.wpsgroovyapi.attributes.DescriptionTypeAttribute
+import org.orbisgis.wpsgroovyapi.attributes.InputAttribute
+import org.orbisgis.wpsgroovyapi.attributes.LiteralDataAttribute
 
 /**
- * Groovy annotation that can be used in a groovy script to declare a literal data input field.
- * This annotation define a field as a literal data.
- * It is used for basic type (boolean, char, String, int double ...)
+ * LiteralData input annotation.
+ * The LiteralData represents a number or a string.
+ * As an input, this annotation should be placed just before the variable.
+ *
+ * The following fields must be defined (mandatory) :
+ *  - title : String
+ *       Title of a process, input, and output. Normally available for display to a human.
+ *
+ * The following fields can be defined (optional) :
+ *  - resume : String
+ *      Brief narrative description of a process, input, and output. Normally available for display to a human..
+ *  - keywords : String
+ *      Coma separated keywords that characterize a process, its inputs, and outputs.
+ *  - identifier : String
+ *      Unambiguous identifier of a process, input, and output. It should be a valid URI.
+ *  - metadata : MetaData[]
+ *      Reference to additional metadata about this item.
+ *  - minOccurs : int
+ *      Minimum number of times that values for this parameter are required. 0 means the input is optional.
+ *  - maxOccurs : int
+ *      Maximum number of times that this parameter may be present.
+ *  - input : InputAttribute[]
+ *      Nested Input.
+ *  - extensions : String[]
+ *      File extension accepted. If no extension are specified, all the format are accepted.
+ *  - isSpatial : boolean
+ *      Indicates if the data represented is spatial or not.
+ *  - isCreateTable : boolean
+ *      Indicates if the toolbox should load the file and give back the table name or just give the file name.
+ *
+ * Usage example can be found at https://github.com/orbisgis/orbisgis/wiki/
  *
  * @author Sylvain PALOMINOS
  */
-@AnnotationCollector([Field, InputAttribute, LiteralDataAttribute, DescriptionTypeAttribute])
+@AnnotationCollector([Field, LiteralDataAttribute, InputAttribute, DescriptionTypeAttribute])
 @interface LiteralDataInput {}
